fix: audit v6 — 62 issues patched (phases 1-8) v3.2.6+18
Phase 1: test coverage gaps (M-001/M-006/M-014/M-013/M-032/M-004)
Phase 2: financial integrity (M-011/M-008/M-009/M-025/M-012)
Phase 3: security (M-024/M-018/M-020/M-016/M-028)
Phase 4: Firestore indexes (already covered)
Phase 5: Android build security (M-002/M-003) — minify+shrink+signing hardened
Phase 6: L10n+UX — about_screen tr(), 3 new keys, dead search_provider deleted
Phase 7: tests 206->272 (sanitizer/error_mapper/invoice_guard/mark_paid)
Phase 8: release — new keystore (BOM bug fixed), split-per-abi APKs + web deployed

+++ app/test/unit/business_logic/void_invoice_balance_test.dart
@@ -180,7 +180,7 @@ void main() {
     test('zero qty entries in rawDeductions are skipped', () {
       final rawDeductions = {'inv-doc-1': 0, 'inv-doc-2': 12};
       final validEntries =
-          rawDeductions.entries.where((e) => (e.value as int) > 0).toList();
+          rawDeductions.entries.where((e) => e.value > 0).toList();
       expect(validEntries.length, equals(1));
       expect(validEntries.first.key, equals('inv-doc-2'));
     });
